Installing the JDK rather than the JRE
A recent blog post described slow import speeds. In
an attempt to get a stacktrace of the hanging code,
we installed the JDK after which it was no longer
possible to reproduce the hang. Installing the JDK
now across the board with the help that this helps.

See: https://www.openmicroscopy.org/community/viewtopic.php?f=5&t=7885